Neighborhood doesn' work

+++ src/queue.c
@@ -2,9 +2,9 @@
 #include "rcm.h"
 
 //! Initializes the Queue.
-Queue* init(size_t size) {
+Queue* init(int size) {
 	Queue *q = malloc(sizeof(Queue));
-	q->buffer = malloc( size * sizeof(unsigned int) );
+	q->buffer = malloc( size * sizeof(int) );
 	q->head = 0;
 	q->tail = 0;
 	q->n = size;
@@ -12,7 +12,7 @@ Queue* init(size_t size) {
 }
 
 //! Adds an element to the queue.
-void push( Queue *Q, unsigned int key ) {
+void push( Queue *Q, int key ) {
 	if( Q->tail < Q->n ) {
 		Q->buffer[Q->tail++] = key;
 	}
